feat: file type specification option added fix: help option reworked fix: small cosmetic changes across the whole program

+++ archivedater.c
@@ -17,6 +17,10 @@
 // Global variable for the log file
 FILE *log_file;
 
+// Global flags to track the creation of temp_dir and temp_files
+int temp_dir_created = 0;
+int temp_files_created = 0;
+
 // Log Levels
 typedef enum {
     DEBUG,
@@ -26,16 +30,16 @@ typedef enum {
 } LogLevel;
 
 // var for current_log_level
-LogLevel current_log_level = INFO;
+LogLevel current_log_level = WARN;
 
 typedef enum {
     OPT_VERBOSE,
     OPT_QUIET,
+    OPT_FILETYPE,
     OPT_LOGLEVEL,
     OPT_OUTPUTDIR,
     OPT_HELP,
     OPT_VERSION,
-    OPT_DRYRUN,
     OPT_UNKNOWN
 } Option;
 
@@ -46,29 +50,35 @@ typedef struct {
 
 
 // Function prototypes
-void process_archive(const char *archive_path, const char *archive_type, const char *file_extension);
-char* read_archive_type(const char *file_path);
+void process_archive(char *archive_path, const char *archive_type, const char *file_extension, char *working_dir);
+char* read_archive_type(const char *archive_path);
 void log_message(LogLevel level, const char *format, ...);
-char* read_file_extension(int argc, char *argv[]);
+char* check_usage(int argc, char *argv[]);
 void create_temp_dir();
-void extract_archive_to_temp_dir(const char *archive_path, const char *archive_type);
+void extract_archive_to_temp_dir(char *archive_path, const char *archive_type);
 void change_file_mod_dates_in_temp_dir(const char *file_extension);
-void get_mod_date(const char *file_extension);
+void get_mod_date();
 char* read_mod_date_from_file(const char *filename);
 void set_mod_dates(const char *file_extension, const char *max_mod_date);
-void create_updated_archive_from_temp_dir(const char *archive_path, const char *archive_type);
+void create_updated_archive_from_temp_dir(char *archive_path, const char *archive_type, char *working_dir);
 void delete_temp_files();
 void remove_temp_dir();
 void execute_command(const char *command);
 void init_log_file();
-void parse_command_line_arguments(int argc, char *argv[]);
+void parse_command_line_arguments(int argc, char *argv[], char **working_dir, char **file_extension);
 ParsedOption parse_option(const char *arg);
 void register_signal_handlers();
 void signal_handler(int signal);
 void register_cleanup_functions();
+void print_help_message();
+char* extract_file_name(char *path);
 void close_log_file();
 
 int main(int argc, char *argv[]) {
+
+    char* working_dir = argv[1];
+    char *file_extension = "*";
+
     // Initialize log file
     init_log_file();
 
@@ -79,20 +89,26 @@ int main(int argc, char *argv[]) {
     register_cleanup_functions();
 
         // Parse command-line arguments
-    parse_command_line_arguments(argc, argv);
-
-    // Read the file extension from command-line arguments
-    char *file_extension = read_file_extension(argc, argv);
+    parse_command_line_arguments(argc, argv, &working_dir, &file_extension);
 
     // Extract and process the archive
-    const char *archive_path = argv[2];
+    char *archive_path = argv[1];
     const char *archive_type = read_archive_type(archive_path);
-    process_archive(archive_path, archive_type, file_extension);
+    process_archive(archive_path, archive_type, file_extension, working_dir);
 
-    // log success message
-    log_message(INFO, "Success");
+    printf("The archive has been updated successfully.\n");
 
     return 0;
+}
+
+// Function for extracting the file name from a path
+char *extract_file_name(char *path) {
+    char *last_slash = strrchr(path, '/');
+    if (last_slash == NULL) {
+        return path; 
+    } else {
+        return last_slash + 1; 
+    }
 }
 
 // Signal handler function
@@ -126,20 +142,20 @@ void register_cleanup_functions() {
 ParsedOption parse_option(const char *arg) {
     ParsedOption result = {OPT_UNKNOWN, NULL};
 
-    if (strcmp(arg, "-v") == 0) {
+    if (strcmp(arg, "-v") == 0 || strcmp(arg, "--verbose") == 0) {
         result.option = OPT_VERBOSE;
-    } else if (strcmp(arg, "-q") == 0) {
+    } else if (strcmp(arg, "-q") == 0 || strcmp(arg, "--quiet") == 0) {
         result.option = OPT_QUIET;
-    } else if (strcmp(arg, "-l") == 0) {
+    } else if (strcmp(arg, "-f") == 0 || strcmp(arg, "--filetype") == 0) {
+        result.option = OPT_FILETYPE;
+    } else if (strcmp(arg, "-l") == 0 || strcmp(arg, "--loglevel") == 0) {
         result.option = OPT_LOGLEVEL;
-    } else if (strcmp(arg, "-o") == 0) {
+    } else if (strcmp(arg, "-o") == 0 || strcmp(arg, "--outputdir") == 0) {
         result.option = OPT_OUTPUTDIR;
-    } else if (strcmp(arg, "-h") == 0) {
+    } else if (strcmp(arg, "-h") == 0 || strcmp(arg, "--help") == 0) {
         result.option = OPT_HELP;
     } else if (strcmp(arg, "--version") == 0) {
         result.option = OPT_VERSION;
-    } else if (strcmp(arg, "-d") == 0) {
-        result.option = OPT_DRYRUN;
     } else {
         result.option = OPT_UNKNOWN;
     }
@@ -149,7 +165,7 @@ ParsedOption parse_option(const char *arg) {
 
 
 // Function for parsing command-line arguments
-void parse_command_line_arguments(int argc, char *argv[]) {
+void parse_command_line_arguments(int argc, char *argv[], char **working_dir, char **file_extension) {
     for (int i = 1; i < argc; i++) {
         ParsedOption option = parse_option(argv[i]);
 
@@ -161,6 +177,15 @@ void parse_command_line_arguments(int argc, char *argv[]) {
             case OPT_QUIET:
                 current_log_level = ERROR;
                 log_message(INFO, "Quiet mode enabled");
+                break;
+            case OPT_FILETYPE:
+                if (i + 1 < argc) {
+                    i++;
+                    *file_extension = argv[i];
+                } else {
+                    log_message(ERROR, "Missing argument for -f");
+                }
+                log_message(INFO, "File type specified");
                 break;
             case OPT_LOGLEVEL:
                 if (i + 1 < argc) {
@@ -189,7 +214,9 @@ void parse_command_line_arguments(int argc, char *argv[]) {
                         perror("chdir() to specified output directory failed");
                         exit(1);
                     }
-                    log_message(INFO, "Changed working directory to %s", argv[i]);
+                    *working_dir = argv[i];
+                    log_message(INFO, "Changed output directory to %s", argv[i]);
+                    log_message(WARN, "The updated archive will be created in the specified output directory");
                 } else {
                     log_message(ERROR, "Missing argument for -o");
                     exit(1);
@@ -216,17 +243,23 @@ void parse_command_line_arguments(int argc, char *argv[]) {
 // Function for printing help message
 void print_help_message() {
     printf("NAME\n");
-    printf("    archivedater - Corrects dates within archives, prioritizing the most relevant file types\n");
+    printf("    archivedater - corrects the dates in the archives by the last file modification date\n");
     printf("\n");
     printf("SYNOPSIS\n");
-    printf("    archivedater [FILE_TYPE] [ARCHIVE_PATH] [OPTIONS]\n");
+    printf("    archivedater [ARCHIVE_PATH] [OPTIONS]\n");
     printf("\n");
     printf("DESCRIPTION\n");
     printf("    This utility is designed to correct dates within archives, prioritizing the most relevant file types specified by the user. It aims to enhance the accuracy and consistency of date information across various archived files, providing a streamlined approach to archive management. By allowing users to select the types of files to be considered for date correction, this utility offers a flexible solution tailored to specific archive needs.\n");
     printf("\n");
-    printf("OPTIONS\n");
+    printf("OPTIONS (no need to specify [ARCHIVE_PATH])\n");
     printf("    -h, --help\n");
     printf("        Display this help message and exit.\n");
+    printf("    --version\n");
+    printf("        Display version information and exit.\n");
+    printf("\n");
+    printf("OPTIONS (need to specify [ARCHIVE_PATH])\n");
+    printf("    -f, --filetype\n");
+    printf("        Set file type to be processed.\n");
     printf("    -v, --verbose\n");
     printf("        Enable verbose mode.\n");
     printf("    -q, --quiet\n");
@@ -235,28 +268,24 @@ void print_help_message() {
     printf("        Set log level (DEBUG, INFO, WARN, ERROR).\n");
     printf("    -o, --outputdir\n");
     printf("        Set output directory.\n");
-    printf("    --version\n");
-    printf("        Display version information and exit.\n");
-    printf("    -d, --dryrun\n");
-    printf("        Perform a dry run without making changes.\n");
     printf("\n");
     printf("EXIT STATUS\n");
     printf("    0 Successful execution.\n");
     printf("    1 An error occurred.\n");
     printf("\n");
     printf("SEE ALSO\n");
-    printf("    man(1), man-pages(7)\n");
+    printf("    https://github.com/Krame1S/KOLYANOV_OSISP_prj_2024/blob/master/README.md\n");
 }
 
 // Function for processing the archive
-void process_archive(const char *archive_path, const char *archive_type, const char *file_extension) {
+void process_archive(char *archive_path, const char *archive_type, const char *file_extension, char *working_dir) {
     create_temp_dir();
     extract_archive_to_temp_dir(archive_path, archive_type);
     change_file_mod_dates_in_temp_dir(file_extension);
-    get_mod_date(file_extension);
+    get_mod_date();
     char *max_mod_date = read_mod_date_from_file("max_mod_date.txt");
     set_mod_dates(file_extension, max_mod_date);
-    create_updated_archive_from_temp_dir(archive_path, archive_type);
+    create_updated_archive_from_temp_dir(archive_path, archive_type, working_dir);
 }
 
 // Logging function with timestamps and log levels
@@ -280,6 +309,7 @@ void log_message(LogLevel level, const char *format, ...) {
     strftime(timestamp, sizeof(timestamp), "%Y-%m-%d %H:%M:%S", timeinfo);
 
     // Log to console
+    usleep(20000);
     printf("%s %s: ", timestamp, level == DEBUG ? "DEBUG" : level == INFO ? "INFO" : level == WARN ? "WARN" : "ERROR");
     vprintf(format, args);
     printf("\n");
@@ -300,27 +330,27 @@ void log_message(LogLevel level, const char *format, ...) {
 
 
 // Function for reading the file extension from command-line arguments
-char* read_archive_type(const char *file_path) {
+char* read_archive_type(const char *archive_path) {
     // Check for .tar.gz, .tar.bz2, or .zip extensions
-    int path_length = strlen(file_path);
-    if (path_length >= 7 && strcmp(file_path + path_length - 7, ".tar.gz") == 0) {
+    int path_length = strlen(archive_path);
+    if (path_length >= 7 && strcmp(archive_path + path_length - 7, ".tar.gz") == 0) {
         return "tar.gz";
-    } else if (path_length >= 8 && strcmp(file_path + path_length - 8, ".tar.bz2") == 0) {
+    } else if (path_length >= 8 && strcmp(archive_path + path_length - 8, ".tar.bz2") == 0) {
         return "tar.bz2";
-    } else if (path_length >= 4 && strcmp(file_path + path_length - 4, ".zip") == 0) {
+    } else if (path_length >= 4 && strcmp(archive_path + path_length - 4, ".zip") == 0) {
         return "zip";
     } else {
-        log_message(ERROR, "Unsupported archive type or no file extension found in '%s'", file_path);
+        log_message(ERROR, "Unsupported archive type or archive path '%s'", archive_path);
         exit(1);
     }
 }
 
 // Function for reading the file extension from command-line arguments
-char* read_file_extension(int argc, char *argv[]) {
+char* check_usage(int argc, char *argv[]) {
     if (argc >= 2) {
         return argv[1];
     } else {
-        log_message(ERROR, "Usage: program file_extension");
+        log_message(ERROR, "Usage: archivedater [ARCHIVE_PATH] [OPTIONS]");
         exit(1);
     }
 }
@@ -331,10 +361,11 @@ void create_temp_dir() {
     snprintf(command, sizeof(command), "mkdir temp_dir");
     execute_command(command);
     log_message(INFO, "Created temporary directory");
+    temp_dir_created = 1;
 }
 
 // Function for extracting an archive to the temporary directory
-void extract_archive_to_temp_dir(const char *archive_path, const char *archive_type) {
+void extract_archive_to_temp_dir(char *archive_path, const char *archive_type) {
     char command[COMMAND_SIZE];
     if (strcmp(archive_type, "tar.gz") == 0) {
         snprintf(command, sizeof(command), "tar -xzvf %s -C temp_dir", archive_path);
@@ -354,14 +385,16 @@ void extract_archive_to_temp_dir(const char *archive_path, const char *archive_t
 void change_file_mod_dates_in_temp_dir(const char *file_extension) {
     char command[COMMAND_SIZE];
     snprintf(command, sizeof(command), "find temp_dir -type f -name '*%s' -printf '%%T@ %%p\\n' | sort -n | tail -n 1 | awk '{print $2}' > max_mod_file.txt", file_extension);
+    temp_files_created = 1;
     execute_command(command);
     log_message(INFO, "Changed file modification dates in temporary directory");
 }
 
 // Function for getting the modification date of a file
-void get_mod_date(const char *file_extension) {
+void get_mod_date() {
     char command[COMMAND_SIZE];
     snprintf(command, sizeof(command), "stat -c %%y $(cat max_mod_file.txt) | awk '{print $1, $2}' > max_mod_date.txt");
+    temp_files_created = 1;
     execute_command(command);
     log_message(INFO, "Got modification date of a file");
 }
@@ -394,18 +427,22 @@ void set_mod_dates(const char *file_extension, const char *max_mod_date) {
     char command[COMMAND_SIZE];
     snprintf(command, sizeof(command), "find temp_dir -name '*%s' -exec touch -d '%s' {} +", file_extension, max_mod_date);
     execute_command(command);
-    log_message(INFO, "Set modification dates for all files of a certain type");
+    log_message(INFO, "Set modification dates for files");
 }
 
 // Function for creating an updated archive from the temporary directory
-void create_updated_archive_from_temp_dir(const char *archive_path, const char *archive_type) {
+void create_updated_archive_from_temp_dir(char *archive_path, const char *archive_type, char *working_dir) {
     char command[COMMAND_SIZE];
+    char *filename = extract_file_name(archive_path);
+    if(archive_path == working_dir)
+        filename = "";
+    archive_path = working_dir;
     if (strcmp(archive_type, "tar.gz") == 0) {
-        snprintf(command, sizeof(command), "tar -czvf %s -C temp_dir .", archive_path);
+        snprintf(command, sizeof(command), "tar -czvf %s%s -C temp_dir .", archive_path, filename);
     } else if (strcmp(archive_type, "tar.bz2") == 0) {
-        snprintf(command, sizeof(command), "tar -cjvf %s -C temp_dir .", archive_path);
+        snprintf(command, sizeof(command), "tar -cjvf %s%s -C temp_dir .", archive_path, filename);
     } else if (strcmp(archive_type, "zip") == 0) {
-        snprintf(command, sizeof(command), "cd temp_dir && zip -r %s .", archive_path);
+        snprintf(command, sizeof(command), "cd temp_dir && zip -r %s%s .", archive_path, filename);
     } else {
         log_message(ERROR, "Unsupported archive type: %s", archive_type);
         exit(1);
@@ -416,18 +453,22 @@ void create_updated_archive_from_temp_dir(const char *archive_path, const char *
 
 // Function for deleting temporary files
 void delete_temp_files() {
-    char command[COMMAND_SIZE];
-    snprintf(command, sizeof(command), "rm -f max_mod_file.txt max_mod_date.txt");
-    execute_command(command);
-    log_message(INFO, "Deleted temporary files");
+    if (temp_files_created) {
+        char command[COMMAND_SIZE];
+        snprintf(command, sizeof(command), "rm -f max_mod_file.txt max_mod_date.txt");
+        execute_command(command);
+        log_message(INFO, "Deleted temporary files");
+    }
 }
 
 // Function for removing the temporary directory
 void remove_temp_dir() {
-    char command[COMMAND_SIZE];
-    snprintf(command, sizeof(command), "rm -rf temp_dir");
-    execute_command(command);
-    log_message(INFO, "Removed temporary directory");
+    if (temp_dir_created) {
+        char command[COMMAND_SIZE];
+        snprintf(command, sizeof(command), "rm -rf temp_dir");
+        execute_command(command);
+        log_message(INFO, "Removed temporary directory");
+    }
 }
 
 // Function for executing a command
